+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 # v2.3.0
 
-This is a maintenance release.
-
-## Fixes
+This is a feature release.
 
  * Fixes a bug in the mock schema registry client where the wrong ID was being
    returned for pre-registered schema (#971, @srlk).
+ * Adds `CreateTopic` method to the MockCluster. (#1047, @mimikwang).
 
 
 # v2.2.0